requirements.txt
PyYAML==6.0.1
click==8.1.7
feedparser==6.0.11
APScheduler==3.10.4
aiofile==3.8.8
httpx[http2]==0.27.2
pydantic==2.9.2
pypinyin==0.53.0
fastapi==0.109.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==11.3.0
numpy==2.3.1
scikit-learn==1.7.0